improvements: wrote platform dependent files so that native bundling is lighter
