+++ src/shortcuts-manager/ShortcutsManager/ShortcutsManager.native.tsx
@@ -1,0 +1,8 @@
+import { memo } from 'react';
+import type { ShortcutsManagerProps } from './utils';
+
+const ShortcutsManager = (_props: ShortcutsManagerProps) => {
+    return null;
+};
+
+export default memo(ShortcutsManager);
