+++ src/shortcuts-manager/ShortcutsManager/index.tsx
@@ -1,0 +1,2 @@
+export { default as ShortcutsManager } from './ShortcutsManager';
+export * from './utils';
