+++ src/shortcuts-manager/index.tsx
@@ -1,4 +1,4 @@
-export { default as ShortcutsManager, Props as ShortcutsManagerProps } from './shortcuts-manager';
-export { default as useShortcut } from './useShortcut';
-export { default as useSetScopes } from './useSetScopes';
+export * from './useSetScopes';
+export * from './ShortcutsManager';
+export * from './useShortcut';
 export * from './types';
