+++ src/shortcuts-manager/useSetScopes/index.tsx
@@ -1,0 +1,1 @@
+export { default as useSetScopes } from './useSetScopes';
