+++ src/shortcuts-manager/useSetScopes/useSetScopes.native.tsx
@@ -1,0 +1,9 @@
+import { useCallback } from 'react';
+
+import type { Scope } from '../types';
+
+const useSetScopes = () => {
+    return useCallback((_newScopes: Scope[] | ((currentScopes: Scope[]) => Scope[])) => {}, []);
+};
+
+export default useSetScopes;
