+++ src/shortcuts-manager/useShortcut/index.tsx
@@ -1,0 +1,1 @@
+export { default as useShortcut } from './useShortcut';
