+++ src/shortcuts-manager/useShortcut/useShortcut.native.tsx
@@ -1,0 +1,5 @@
+import type { Shortcut } from '../types';
+
+const useShortcut = (_name: string, _callback: (shortcut: Shortcut) => void) => {};
+
+export default useShortcut;
